Feature/pypz 5 amqp plugin (#12)
* Added AMQP plugin implementation

* Reverted interrupt handling in abstract channel rw plugins, since it caused infinite loop.

* Added feature to provide error source and exception to the _on_error method; Channel resources will not be deleted on error from operation states;

* Fixed bug: connections not closed on exception in open(); Fixed bug: instance has been checked in abstract channel ports' _on_error instead of class;

* Added RMQ docs

* Added Avro serde support for RMQ

+++ sniffer/src/pypz/sniffer/viewer.py
@@ -53,7 +53,7 @@ class PipelineSnifferViewer(tk.Tk):
         self._operator_view_channel_writer_count: dict[Operator, int] = dict()
         self._stopped: bool = False
         self._shutdown: bool = False
-        self._start_time: int = 0
+        self._start_time: float = 0.0
 
         # Initialize graphical elements
         # =============================
